refactor: move formats to src/formats/ module with unified TensorView trait
- Moved safetensors/ and serverlessllm/ under src/formats/
- Moved error.rs and traits.rs into src/formats/ (no separate types/ module)
- Added TensorView trait for uniform tensor access across formats
- Changed serverlessllm to use usize instead of i64 for shapes/strides
- Updated all import paths in converters, bin files, and profiling code

+++ src/formats/serverlessllm/types.rs
@@ -17,10 +17,10 @@ pub struct TensorEntry {
     pub size: u64,
 
     /// Shape of the tensor (dimensions)
-    pub shape: Vec<i64>,
+    pub shape: Vec<usize>,
 
     /// Stride information for the tensor
-    pub stride: Vec<i64>,
+    pub stride: Vec<usize>,
 
     /// Data type of the tensor (e.g., "float32", "int64")
     pub dtype: String,
